Add dhh_auth app with custom user model
This includes a new permission model where there can be more then one user class
in the system. There will still be one "main" user class but other user classes can be
derived from the AbstractClientUser.

Instead of relying on Djangos build in permissions system, there is a new ClientPermission
system in place. This is because Djangos default permission system only supports
permissions for the main user model. The permission system is now moved to a
client, which can be included in multiple user classes. The ClientPermission
model still follows Djangos default permission model when it comes to structure.

A custom token authentication base class is created that can be extended to accomodate
for multiple authentication methods while still using the Client and ClientPermission
model. Normally only a (<user>, None) is returned from the authenticate method
but we instead return (<user>, <client>).

The reason for doing this is so that we can have multiple types of users. Right now this
is mainly for authenticating services in a similar fashion as normal users, but can in
the future be used for other user classes as well.