Add service app
Works in a similar way as the dhh_auth app in that it creates a new
user model, Service, which supports Client and ClientPermission.

It also adds endpoints for creating service, service tokens and
service permissions.

Service permissions are permissions given by a user to a service
for a specific Datastream. Right now the permissions are given to
the internal ID of a Datastream Django model instance. This is
due to the way ClientPermissions work so that generic foreign
keys work. This may cause a bit of confusion since the id of the
Datastream on the Django side is never exposed currently. This
is something that we can fix in a later version by rewriting parts
of the serializer to accept STS IDs and the translate that to an
internal Datastream id.